+++ cocktagorize/src/main/java/com/ossp/cocktagorize/service/CocktailService.java
@@ -11,6 +11,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Service
@@ -70,7 +71,7 @@ public class CocktailService {
         return cocktailRepository.findByTagsByOr(tags, pageable).map(CocktailResponseDto::new);
     }
 
-    public Page<CocktailResponseDto> getCocktailByDegree(int min,int max,Authentication authentication,Pageable pageable){
+    public Page<CocktailResponseDto> getCocktailByDegree(BigDecimal min, BigDecimal max, Authentication authentication, Pageable pageable){
         if(authentication != null && authentication.getPrincipal() != "anonymousUser"){
             return cocktailRepository.findByAlcoholDegreeBetweenOrderByAlcoholDegreeAsc(min,max,pageable).map(cocktail -> new CocktailResponseDto(cocktail,userLikeCocktailRepository.findByCocktailIdAndUserId(cocktail.getId(),userRepository.findByUsername(authentication.getName()).getId())));
         }
